initial: koji-image-builder
This is the initial work in progress implementation of
`koji-image-builder`. A `koji` plugin that allows building images with
`image-builder`.

It's always a bit arbitrary when a project gets its first commit but in
this case everything works "on my pc" and it is ready to start iterating
on with integration tests, documentation, and fixes.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>