test: setup a layout for tests and local env
Sets up `run.py`, a script that sets up a local `koji` environment and
can run a smoke test for building an image with the `koji-image-builder`
plugin.

Also sets up initial test files and adds initial test dependencies to
specfiles.

The integration test setup is copied from tkopecek's [1], it is adjusted
to ensure that the necessary components rootful where necessary so that
image builds succeed.

Certificates for the full setup are included, we don't regenerate these
between runs so don't expose any of this publicly as anyone can generate
certificates to connect to your instance.

[1]: https://github.com/tkopecek/koji-container-dev

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>