generic: allow multiple types
`types` is a list of image types to build, when we receive multiple
types we build in a loop inside the same buildroot. This allows
`image-builder` to make use of caching since filesystem trees are often
shared, and so are RPMs.

The cache is per-task.

Note that `pungi`, which does composes, doesn't quite understand it if
one build produces multiple artifacts. This is fine, it could be added
in the future and in the meantime this can be used in scratch builds.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>