ci: enable epel8 in packit
We'll need to ship this package in epel8 as well as CentOS stream
infrastructure runs on 8. Adjust the PackIt config to do so.

I've already done scratch builds and a manual release to epel8 of the
latest version but we might want to to a bump after this PR is merged to
have PackIt handle the branches.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>