many: allow `bootc`
`image-builder >= 34` contains experimental support for building a
subset of image types from bootable containers. Let's allow passing
these options along through the plugin.

Using this option will mean that the builder needs access to where the
bootable container reference is to be pulled from. It also implies that
`podman` is installed in the build root.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>